backend/foodgram/requirements.txt
Django==3.2
djangorestframework
djoser
drf-extra-fields
django_filter
python-dotenv
Pillow
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
django_extensions